[DF-11715] build fixed
Former-commit-id: 0dcc1e6ef90c74137895f48502323f0dd23f5dcd

+++ Modules/QtHelpers/QtHelpers/Private/LauncherListener.cpp
@@ -50,7 +50,7 @@ void LauncherListener::OnReadyRead()
     }
     QByteArray data = clientConnection->readAll();
     bool ok = false;
-    long code = data.toLong(&ok);
+    int code = data.toInt(&ok);
     LauncherMessageCodes::eReplyInternal replyInternal = LauncherMessageCodes::eReplyInternal::WRONG_MESSAGE_FORMAT;
     if (ok)
     {
@@ -65,7 +65,7 @@ void LauncherListener::OnReadyRead()
             replyInternal = static_cast<LauncherMessageCodes::eReplyInternal>(reply);
         }
     }
-    QByteArray reply = QByteArray::number(static_cast<long>(replyInternal));
+    QByteArray reply = QByteArray::number(static_cast<int>(replyInternal));
     clientConnection->write(reply);
     clientConnection->flush();
     clientConnection->disconnectFromServer();
